Change audio recorder to use 16000 sample rate
This enables easier use of a local whispercpp server should it ever be included within wingman itself or if a user wants to use a whispercpp server instead of using azure or openai paid cloud services.

When making this change to other repos in the past I have seen no downsides. File size also remains smaller this way.