Add get_date_time_skill (#258)
Add a simple skill to be able to tell the AI what date and time it is. Usually the user won't use this skill directly (though perhaps it could come in handy in VR where the user can't see the desktop) but it would instead typically be called by wingman in chained command fashion so if the user references something that happened "yesterday" or "last night" or "last week" the AI will first look up the date and time to respond accordingly.  Otherwise the AI tends to think "today" or "this year" is sometime in 2023, or whenever its last training data was.